+++ pyincore_viz/geoutil.py
@@ -6,6 +6,7 @@
 
 from pathlib import Path
 
+import os
 import contextily as ctx
 import geopandas as gpd
 import ipyleaflet as ipylft
@@ -15,11 +16,14 @@ import rasterio
 import rasterio.plot
 
 from pyincore import Dataset
+from pyincore import NetworkDataset
 from pyincore.dataservice import DataService
 from pyincore.hazardservice import HazardService
 from pyincore_viz import globals
 from owslib.wms import WebMapService
 from pyincore_viz.csvmaputil import CsvMapUtil
+
+logger = globals.LOGGER
 
 
 class GeoUtil:
@@ -255,7 +259,7 @@ class GeoUtil:
         Args:
             datasets (list): list of pyincore Dataset objects
             zoom_level (int): initial zoom level
-            wmr_url (str): URL of WMS server
+            wms_url (str): URL of WMS server
             layer_check (bool): boolean for checking the layer availability in wms server
 
         Returns:
@@ -317,7 +321,7 @@ class GeoUtil:
             datasets (list): list of pyincore Dataset objects
             wms_datasets (list): list of pyincore Dataset objects for wms layers
             zoom_level (int): initial zoom level
-            wmr_url (str): URL of WMS server
+            wms_url (str): URL of WMS server
 
         Returns:
             obj: A ipylfealet Map object
@@ -381,3 +385,54 @@ class GeoUtil:
         csvmap = CsvMapUtil.generate_map_csv_from_dir(inventory_dataset, column, file_path)
 
         return csvmap
+
+    @staticmethod
+    def plot_network_dataset(network_dataset: NetworkDataset, zoom_level=10):
+        """Creates map window with Network Dataset visualized
+
+        Args:
+            network_dataset (NetworkDataset):  pyincore Network Dataset object
+            zoom_level (int): zoom level indicator value for mapping
+
+        Returns:
+            m (ipyleaflet.Map): ipyleaflet Map object
+
+        """
+        # get node file name path
+        link_path = network_dataset.link.file_path
+        link_file_name = os.path.basename(link_path)
+
+        # get node file name path
+        node_path = network_dataset.node.file_path
+        node_file_name = os.path.basename(node_path)
+
+        # read file using geopandas
+        node_gdf = gpd.read_file(node_path)
+        link_gdf = gpd.read_file(link_path)
+
+        geo_data_list = []
+        # (min_lat, min_lon, max_lat, max_lon)
+        bbox_all = [9999, 9999, -9999, -9999]
+
+        # add node data to list
+        node_geo_data = ipylft.GeoData(geo_dataframe=node_gdf, name=node_file_name)
+        geo_data_list.append(node_geo_data)
+
+        # add link data to list
+        link_geo_data = ipylft.GeoData(geo_dataframe=link_gdf, name=link_file_name)
+        geo_data_list.append(link_geo_data)
+
+        bbox = link_gdf.total_bounds
+        bbox_all = GeoUtil.merge_bbox(bbox_all, bbox)
+
+        cen_lat, cen_lon = (bbox_all[2] + bbox_all[0]) / 2.0, (bbox_all[3] + bbox_all[1]) / 2.0
+
+        # TODO: ipylft doesn't have fit bound methods, we need to find a way to zoom level to show all data
+        m = ipylft.Map(center=(cen_lon, cen_lat), zoom=zoom_level, basemap=ipylft.basemaps.Stamen.Toner, crs='EPSG3857',
+                       scroll_wheel_zoom=True)
+        for entry in geo_data_list:
+            m.add_layer(entry)
+
+        m.add_control(ipylft.LayersControl())
+
+        return m