+++ tests/pyincore_viz/test_pyincore_viz.py
@@ -1,13 +1,14 @@
 import os
-# Connect to IN-CORE Services
+
 from pyincore import IncoreClient
 from pyincore.dataservice import DataService
 from pyincore.hazardservice import HazardService
 from pyincore_viz.geoutil import GeoUtil as viz
 from pyincore import Dataset
-
+from pyincore import NetworkDataset
 from pyincore.globals import INCORE_API_DEV_URL
 
+# Connect to IN-CORE Services
 client = IncoreClient(INCORE_API_DEV_URL)
 
 # testing datasets
@@ -16,6 +17,7 @@ joplin_bldg_inv_id = "5df7d0de425e0b00092d0082"
 eq_hazard_id = "5b902cb273c3371e1236b36b"
 shelby_hopital_inv_id = "5a284f0bc7d30d13bc081a28"
 shelby_road_id = "5a284f2bc7d30d13bc081eb6"
+centerville_epn_network_id = "5d25fb355648c40482a80e1c"
 
 # Creating a Dataset object with ID and Data Service
 eq_metadata = HazardService(client).get_earthquake_hazard_metadata(eq_hazard_id)
@@ -36,6 +38,11 @@ viz.plot_map(sh_bldg_inv, column="struct_typ", category=False, basemap=True)
 viz.get_wms_map([sh_bldg_inv, sh_road], zoom_level=10)
 viz.get_gdf_map([sh_bldg_inv, sh_road], zoom_level=10)
 viz.get_gdf_wms_map([sh_bldg_inv], [sh_road], zoom_level=10)
+
+# visualize network dataset
+dataset = Dataset.from_data_service(centerville_epn_network_id, DataService(client))
+network_dataset = NetworkDataset(dataset)
+viz.plot_network_dataset(network_dataset, 12)
 
 # visualize tornado
 viz.plot_tornado(tornado_hazard_id, client, basemap=False)